// src/DevIO.API/wwwroot/app/demo-webapi/src/app/corridas/models/Corrida.ts
export class Corrida {
  veiculoId: string;
  idMotoristaPrimeiro: string;
  idMotoristaSegundo: string;
  descricao: string;
  dataHoraSaida: string;
  dataHoraChegada: string;
  tipoViagem: string;
  nomePlaca: string;
  nomeModelo: string;
  nomeMotoristaPrimeiro: string;
  nomeMotoristaSegundo: string;
  documentoMotoristaPrimeiro: string;
  documentoMotoristaSegundo: string;
}